Rename resampled.py to resample_result.py for clarity
Disambiguate the confusingly similar resample.py and resampled.py
file names by renaming the wrapper class module to resample_result.py.
