+++ python/tempo/__init__.py
@@ -1,3 +1,3 @@
-from tempo.resampled import ResampledTSDF  # noqa: F401
+from tempo.resample_result import ResampledTSDF  # noqa: F401
 from tempo.tsdf import TSDF  # noqa: F401
 from tempo.utils import display  # noqa: F401
